// lib/contracts/types.ts
export enum CONTRACTS_TYPE {
  FACTORY = "MaverickV2Factory",
  POOL_LENS = "MaverickV2PoolLens",
  POSITION = "MaverickV2Position",
  ROUTER = "MaverickV2Router",
  QUOTER = "MaverickV2Quoter",
  LIQUIDITY_MANAGER = "MaverickV2LiquidityManager",
  VOTING_DISTRIBUTOR = "VotingDistributor",
  VOTING_ESCROW_NFT = "VotingEscrowNft",
  POOL_2 = "Pool2",
  ROOSTER = "Rooster",
  POOL_DISTRIBUTOR = "PoolDistributor",
  LP_REWARD = "LpReward",
}